apps/infra/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "infra" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/infra/html' --cov-report xml:'../../coverage/apps/infra/coverage.xml' --html='../../reports/apps/infra/unittests/html/index.html' --junitxml='../../reports/apps/infra/unittests/junit.xml'"

[tool.poetry]
name = "infra"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "infra"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  aws-cdk-lib = "^2.147.2"
  constructs = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
